% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_tools.R
\name{as.unordered}
\alias{as.unordered}
\title{Unorder ordered factor}
\usage{
as.unordered(x)
}
\arguments{
\item{x}{Ordered factor}
}
\value{
Unordered factor
}
\description{
Unorder ordered factor
}
\examples{
as.unordered(ordered(c("a","a","b","c")))
}
