% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_utils.R
\name{beta_diff}
\alias{beta_diff}
\title{Closed form solution for difference between two beta distribution}
\usage{
beta_diff(
  event_a,
  n_a,
  event_b,
  n_b,
  prior = c("uniform", "jeffrey", "empirical", "half_empirical"),
  sample = NULL
)
}
\arguments{
\item{event_a}{number of events in group a}

\item{n_a}{size group a}

\item{event_b}{number of events in group b}

\item{n_b}{size group b}

\item{prior}{prior for beta distribution, either 'uniform','jeffrey' or 'emprirical'}

\item{sample}{size of samples}
}
\value{
Posterior probaiblity of superiority
}
\description{
Closed form solution for difference between two beta distribution
}
\examples{
beta_diff(10,11,1,11)
}
\references{
exact solution function, based on https://www.evanmiller.org/bayesian-ab-testing.html#binary_ab
}
