% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement.R
\name{cohenkappa}
\alias{cohenkappa}
\title{Cohen's kappa}
\usage{
cohenkappa(x, weights = c("linear", "quadratic"))
}
\arguments{
\item{x}{An n*n matrix/contingency table}

\item{weights}{A character string specifying "linear" or "quadratic" weights, or a numeric matrix of n*n weights.}
}
\value{
An object of type "kappa"
}
\description{
Cohen's kappa
}
\examples{
x <- matrix(c(106, 10,4,22,28, 10,2, 12,  6),ncol=3,byrow=TRUE)
kappa(x,weights="quadratic")


}
