% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyplots.R
\name{distplot}
\alias{distplot}
\title{Distplot}
\usage{
distplot(.data, x, y, ul = 0, ll = 0, ...)
}
\arguments{
\item{.data}{dataframe to plot}

\item{x}{value}

\item{y}{group}

\item{ul}{upper limit}

\item{ll}{lower limit}

\item{...}{Additional parameters}
}
\value{
A grouped tidybayes halfeye plot coloured according to value
}
\description{
Distplot
}
\examples{
\donttest{
library(tidyverse)

df <- tibble(a = rnorm(1e4, 0, 1), b = rnorm(1e4, 0.25, 1)) \%>\%
  pivot_longer(cols = everything(), names_to = "y", values_to = "x")
df \%>\% distplot(x = x, y = y, ul = 0.1, ll = -0.1)
}
}
