% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_utils.R
\name{h}
\alias{h}
\title{Exact solution for difference between two beta distribution}
\usage{
h(alpha_a, beta_a, alpha_b, beta_b)
}
\arguments{
\item{beta_b}{}
}
\value{

}
\description{
Exact solution for difference between two beta distribution
}
\references{
exact solution function, based on "Introduction to empirical Bayes", Robinson
}
