% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_utils.R
\name{norm_logit_solve}
\alias{norm_logit_solve}
\title{Given mu and value, solve for SD so that the tail area of the normal distribution beyond value is prob}
\usage{
norm_logit_solve(mu, value, prob)
}
\arguments{
\item{mu}{mean value}

\item{value}{value for tail area}

\item{prob}{probability of tail area}
}
\value{

}
\description{
Given mu and value, solve for SD so that the tail area of the normal distribution beyond value is prob
}
\examples{
norm_logit_solve(qlogis(0.05), qlogis(0.2), 0.1) # qlogis is R logit ()
}
\references{
http://hbiostat.org/doc/bbr.pdf
}
