% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_utils.R
\name{normalbeta}
\alias{normalbeta}
\title{Beta distribution parameters from normal distribution mu and variance}
\usage{
normalbeta(mu, sd)
}
\arguments{
\item{mu}{Mean for normal distribution}

\item{sd}{Standard deviation of normal distribution}
}
\value{
list of alpha and beta parameters for beta distribution
}
\description{
Beta distribution parameters from normal distribution mu and variance
}
\examples{
normalbeta(0,1)
}
