% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posterior}
\alias{posterior}
\title{Posterior draws from model m12.1 in Statistical Rethinking}
\format{
A dataframe of posterior draws
}
\usage{
posterior
}
\description{
Posterior draws from model m12.1 in Statistical Rethinking
}
\references{
McElreath (2020). Statistical Rethinking: A Bayesian Course with Examples in R and Stan (2nd edition). CRC Press.
}
\keyword{datasets}
