% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posterior_clean}
\alias{posterior_clean}
\title{Posterior draws renamed with tidynamer}
\format{
A dataframe of posterior drawsfrom m12. with renamed coefficients using tidynamer
}
\usage{
posterior_clean
}
\description{
Posterior draws renamed with tidynamer
}
\references{
McElreath (2020). Statistical Rethinking: A Bayesian Course with Examples in R and Stan (2nd edition). CRC Press.
}
\keyword{datasets}
