% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laRs_methods.R
\name{print.laRs}
\alias{print.laRs}
\title{Print methods}
\usage{
\method{print}{laRs}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of type cohenkappa}

\item{digits}{Rounding}

\item{...}{Additional arguemnts for print}
}
\description{
Print methods
}
