% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_tools.R
\name{snip}
\alias{snip}
\title{Snip carbonate, pretty twitter code
https://twitter.com/jeremy_data/status/1371171913578520587}
\usage{
snip()
}
\description{
Snip carbonate, pretty twitter code
https://twitter.com/jeremy_data/status/1371171913578520587
}
