% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_cmdstanr.R
\name{tidy_cmdstanr}
\alias{tidy_cmdstanr}
\title{Change the structure of cmdstan output to fit tidybayes}
\usage{
tidy_cmdstanr(df)
}
\arguments{
\item{df}{dataframe from cmdstanr to be tidyformed}
}
\value{
data.frame
}
\description{
Change the structure of cmdstan output to fit tidybayes
}
\examples{

posterior \%>\% tidy_cmdstanr()
}
