% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidynamer.R
\name{tidynamer}
\alias{tidynamer}
\title{Change the structure of cmdstan output to fit tidybayes}
\usage{
tidynamer(df)
}
\arguments{
\item{df}{dataframe from cmdstanr to be tidyformed}
}
\value{
data.frame
}
\description{
Change the structure of cmdstan output to fit tidybayes
}
\examples{
data(posterior)
posterior \%>\% tidynamer()
}
