% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{weibull_hazard}
\alias{weibull_hazard}
\title{Weibull hazard function}
\usage{
weibull_hazard(scale, shape, time)
}
\arguments{
\item{scale}{The scale parameter for the Weibull distribution}

\item{shape}{The shape parameter for the Weibull distribution}

\item{time}{Time point or sequence}
}
\value{
hazard
}
\description{
Weibull hazard function
}
\examples{
weibull_hazard(50,0.5,1:100)
}
