% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{weibull_mean}
\alias{weibull_mean}
\title{Weibull mean}
\usage{
weibull_mean(scale, shape)
}
\arguments{
\item{scale}{The scale parameter for the Weibull distribution}

\item{shape}{The shape parameter for the Weibull distribution}
}
\value{
Mean for the weibull distribution
}
\description{
Weibull mean
}
\examples{
weibull_mean(1200,0.5)

}
