% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{weibull_median}
\alias{weibull_median}
\title{Weibull median}
\usage{
weibull_median(scale, shape)
}
\arguments{
\item{scale}{The scale parameter for the Weibull distribution}

\item{shape}{The shape parameter for the Weibull distribution}
}
\value{
Median for the Weibull distribution
}
\description{
Weibull median
}
\examples{
weibull_median(1200, 0.5)
}
