% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{weibull_survival}
\alias{weibull_survival}
\title{Weibull survival function}
\usage{
weibull_survival(scale, shape, time)
}
\arguments{
\item{scale}{The scale parameter for the Weibull distribution}

\item{shape}{The shape parameter for the Weibull distribution}

\item{time}{Time point or sequence}
}
\value{
survival probabilities
}
\description{
Weibull survival function
}
\examples{
weibull_survival(50,0.5,1:100)
}
