% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_utils.R
\name{beta_diff}
\alias{beta_diff}
\title{Closed form solution for difference between two beta distribution}
\usage{
beta_diff(alpha_a, beta_a, alpha_b, beta_b)
}
\arguments{
\item{alpha_a}{alpha parameter distribution a}

\item{beta_a}{beta parameter distribution a}

\item{alpha_b}{alpha parameter distribution b}

\item{beta_b}{beta parameter distribution b}
}
\value{

}
\description{
Closed form solution for difference between two beta distribution
}
\examples{
beta_diff(10,1,1,10)
}
\references{
exact solution function, based on https://www.evanmiller.org/bayesian-ab-testing.html#binary_ab
}
